Move CodePaths and FileCodeSnippet to common directory
