+++ extensions/ql-vscode/src/stories/common/CodePaths.stories.tsx
@@ -3,7 +3,7 @@ import React from 'react';
 import { ComponentStory, ComponentMeta } from '@storybook/react';
 import { ThemeProvider } from '@primer/react';
 
-import CodePaths from '../../view/remote-queries/CodePaths';
+import { CodePaths } from '../../view/common';
 import type { CodeFlow } from '../../remote-queries/shared/analysis-result';
 
 export default {
@@ -112,8 +112,8 @@ PowerShell.args = {
   message: {
     tokens: [
       {
-        type: 'text',
-        t: 'This zip file may have a dangerous path'
+        t: 'text',
+        text: 'This zip file may have a dangerous path'
       }
     ]
   },
