+++ extensions/ql-vscode/src/view/common/CodePaths/index.ts
@@ -1,0 +1,1 @@
+export * from './CodePaths';
