+++ extensions/ql-vscode/src/view/common/FileCodeSnippet/index.ts
@@ -1,0 +1,1 @@
+export * from './FileCodeSnippet';
