+++ extensions/ql-vscode/src/view/common/index.ts
@@ -1,4 +1,6 @@
 export * from './icon';
+export * from './CodePaths';
+export * from './FileCodeSnippet';
 export * from './HorizontalSpace';
 export * from './SectionTitle';
 export * from './VerticalSpace';
