+++ extensions/ql-vscode/src/view/remote-queries/AnalysisAlertResult.tsx
@@ -1,7 +1,6 @@
 import * as React from 'react';
 import { AnalysisAlert } from '../../remote-queries/shared/analysis-result';
-import CodePaths from './CodePaths';
-import FileCodeSnippet from './FileCodeSnippet';
+import { CodePaths, FileCodeSnippet } from '../common';
 
 const AnalysisAlertResult = ({ alert }: { alert: AnalysisAlert }) => {
   const showPathsLink = alert.codeFlows.length > 0;
